remove doesPathExist check in checkKeyCheckSum
Before commit c565340adc1e98886ce8689814f832561c8418a1,
it was statting the file in order to get its size, which was needed to
use an external hasher. In that commit since it no longer needed the
stat, I made it check doesPathExist since the old code implicitly
checked if the file existed. But that is not necessary, this should only
be ever called on files that exist.